refactor: Code block component
