+++ design-system/components/block-code/block-code.types.tsx
@@ -10,19 +10,6 @@ export type BlockCodeProps = {
   css?: SystemStyleObject
 }
 
-export type PreChildrenProps = {
-  props: {
-    className: string
-    children: string
-  }
-}
-
-export type PreProps = {
-  children?: React.ReactNode
-  title?: string
-  highlight?: string | undefined
-}
-
 export type PreElementProps = ComponentPropsWithoutRef<'pre'>
 
 export type CodeElementProps = ComponentPropsWithoutRef<'code'>
