+++ design-system/components/block-code/index.tsx
@@ -1,3 +1,2 @@
 export { BlockCode } from './block-code'
-export { preToCodeProps } from './block-code.helpers'
-export { type PreChildrenProps, type PreProps } from './block-code.types'
+export { isTerminal } from './block-code.helpers'
