+++ types/code.tsx
@@ -1,0 +1,12 @@
+export type PreChildrenProps = {
+  props: {
+    className: string
+    children: string
+  }
+}
+
+export type PreProps = {
+  children?: React.ReactNode
+  title?: string
+  highlight?: string | undefined
+}
